Add methods `int2Frac` for converting int to Fraction. Add methods `Frac2Rad` `int2Rad` for converting Fraction or int to Radical. Add struct Str. Add `toStrRad` `toStrFrac` to convert Radical or Fraction to string (through struct Str). Add `squareRad`, As the name suggests, the square of Radical, returns a Fraction. Enhanced compatibility with Clang. Add an example "simple_vector_calcs.c" which does some simple computation of space vectors (not the container of the similar name in the STL).
Incompatible:
Changed the function of `printRad`, the old `printRad` was renamed as `pprintRad`, and the new `printRad` is a single-line output method. `pprintRad` is buggy and not recommended.
Fix a `gcd_frac` bug which returns negative number.
Improve the behavior when reducing to 0 in `addRad`, it will now retain a 0 instead of becoming empty, to distinguish between 0 and empty, and avoid many troubles. But `initPoly` behavior is unchanged, it is still empty after initialization instead of 0.